backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
torch>=2.0.0
torchaudio>=2.0.0
audiocraft==1.3.0
pydub==0.25.1
openai==1.3.0
python-dotenv==1.0.0
numpy>=1.24.0
librosa>=0.10.0
soundfile>=0.12.0
scipy>=1.10.0
mutagen>=1.47.0